requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
langchain-core
sentence-transformers